requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-tinymce==4.0.0
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.6.0
